+++ lib/jazz_hands/railtie.rb
@@ -4,6 +4,7 @@ require 'pry-git'
 require 'pry-remote'
 require 'pry-nav'
 require 'awesome_print'
+require 'jazz_hands/hirb_ext'
 
 if RUBY_VERSION >= '1.9.3'
   require 'coolline'